+++ src/positionsensing.cpp
@@ -5,17 +5,17 @@ using namespace vex;
 class PositionSensing{
     //Note: All measurements are in inches
     public:
-    std::vector<double> currentPosition;
-    double previousGlobalOrientation;
-    double lastResetGlobalOrientation;
+    std::vector<float> currentPosition;
+    float previousGlobalOrientation;
+    float lastResetGlobalOrientation;
 
     //Left-right distance between tracking center and left tracking wheel
-    double leftTrackingWheelDistance;
+    float leftTrackingWheelDistance;
     //Left-right distance between tracking center and right tracking wheel
-    double rightTrackingWheelDistance;
+    float rightTrackingWheelDistance;
     //Forward-backward distance between tracking center to the back tracking wheel
-    double backTrackingWheelDistance;
-    PositionSensing(double startingX, double startingY, double _leftTrackingWheelDistance, double _rightTrackingWheelDistance, double _backTrackingWheelDistance, double startingOrientation)
+    float backTrackingWheelDistance;
+    PositionSensing(float startingX, float startingY, float _leftTrackingWheelDistance, float _rightTrackingWheelDistance, float _backTrackingWheelDistance, float startingOrientation)
     {
         currentPosition = {startingX, startingY};
         leftTrackingWheelDistance = _leftTrackingWheelDistance;
@@ -24,23 +24,23 @@ class PositionSensing{
         lastResetGlobalOrientation = startingOrientation;
     }
 
-    public: std::vector<double> GetPosition()
+    public: std::vector<float> GetPosition()
     {
         return currentPosition;
     }
 
-    public: void UpdatePosition(double angleLeft, double angleRight, double angleBack)
+    public: void UpdatePosition(float angleLeft, float angleRight, float angleBack)
     {
         //Deltas correspond to the change in angles from the rotation sensors
         //Need to change deltas to distance traveled 
-        double deltaLeft = VectorMath::AngleToDistance(angleLeft);
-        double deltaRight = VectorMath::AngleToDistance(angleRight);
-        double deltaBack = VectorMath::AngleToDistance(angleBack);
+        float deltaLeft = VectorMath::AngleToDistance(angleLeft);
+        float deltaRight = VectorMath::AngleToDistance(angleRight);
+        float deltaBack = VectorMath::AngleToDistance(angleBack);
         //Note arc angle is in radians
-        double deltaTheta = (deltaLeft - deltaRight) / (leftTrackingWheelDistance + rightTrackingWheelDistance);
-        double arcRadius = (deltaRight / deltaTheta) + rightTrackingWheelDistance;
+        float deltaTheta = (deltaLeft - deltaRight) / (leftTrackingWheelDistance + rightTrackingWheelDistance);
+        float arcRadius = (deltaRight / deltaTheta) + rightTrackingWheelDistance;
 
-        std::vector<double> globalTranslationVector = {(deltaBack / deltaTheta) + backTrackingWheelDistance, (deltaRight / deltaTheta) + rightTrackingWheelDistance};
+        std::vector<float> globalTranslationVector = {(deltaBack / deltaTheta) + backTrackingWheelDistance, (deltaRight / deltaTheta) + rightTrackingWheelDistance};
         //Have to multiply vector by 2sin(theta/2) t
         VectorMath::ScaleVector(globalTranslationVector, 2 * sin(deltaTheta / 2.0));
         VectorMath::CartesianToPolar(globalTranslationVector);
